pyproject.toml
[project]
name = "amptudix-research-agent-native"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "ddgs>=9.11.4",
    "httpx>=0.28.1",
    "openai>=2.28.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "respx>=0.22.0",
    "rich>=14.3.3",
    "trafilatura>=2.0.0",
]
